backend/requirements.txt

# --- FastAPI & Server ---
fastapi==0.115.0
uvicorn[standard]==0.30.6

# --- Database ---
SQLAlchemy==2.0.34
psycopg2-binary==2.9.9

# --- Supabase ---
supabase==2.3.0

# --- Auth & Security ---
python-jose==3.3.0
passlib[bcrypt]==1.7.4
pydantic[email]==2.9.2
python-multipart==0.0.9

# --- PDF Processing ---
PyPDF2==3.0.1

# --- OpenAI / LLM ---
openai==1.50.2

# --- Background Tasks ---
celery==5.3.6
redis==5.0.1

